선호도 증가 알고리즘 구현

+++ src/main/java/com/example/demo/model/UserPreference.java
@@ -24,7 +24,8 @@ public class UserPreference {
     @Column
     private int applicationId;
 
+    //UserRecommendApiController의 dataToCSV의 로직에 의해 int를 float으로 바꿈
     @Column
-    private int preference;
+    private float preference;
 
 }
